Moved db interfaces into db package and converted Track to interface
