--- app/src/main/java/classes/Track.kt
@@ -1,3 +1,0 @@
-package classes
-
-data class Track(val name: String, val author: String, val id: Int)
