+++ app/src/main/java/com/chebureck/playlist/db/Track.kt
@@ -1,0 +1,7 @@
+package com.chebureck.playlist.db
+
+interface Track {
+    val name: String
+    val author: String
+    val id: Int
+}
